Refactor Events listings to separate non-flux component
